Explicit int cast from size_t.
git-svn-id: https://omniorb.svn.sourceforge.net/svnroot/omniorb/branches/4_1@5972 31ce16a7-c309-44a6-b7c4-d857baf782c7

+++ omniORB/include/omniORB4/callDescriptor.h
@@ -322,7 +322,7 @@ public:
     inline _cCORBA_mObject_i_cstring(LocalCallFn lcfn, const char* op_,
 		     size_t oplen, const char* a_0,
 		     _CORBA_Boolean upcall=0) :
-      omniCallDescriptor(lcfn, op_, oplen, 0, 0, 0, upcall),
+      omniCallDescriptor(lcfn, op_, (int)oplen, 0, 0, 0, upcall),
       arg_0( OMNI_CONST_CAST(char*, a_0) ) {}
  
     inline ~_cCORBA_mObject_i_cstring() {
